requirements.txt
bedrock-agentcore
langgraph
langchain-core
langchain-aws
langchain-google-genai
typing-extensions
python-dotenv
requests
